+++ example/gcp/provider.tf
@@ -1,0 +1,14 @@
+terraform {
+  required_providers {
+    conformity = {
+      version = "0.3.6"
+      source  = "trendmicro/conformity"
+
+    }
+  }
+}
+
+provider "conformity" {
+  region = var.region
+  apikey = var.apikey
+}