+++ example/gcp/variable.tf
@@ -1,0 +1,29 @@
+variable "apikey"{
+    type    = string
+    default = ""
+}
+variable "region"{
+    type    = string
+    default = ""
+}
+
+variable "gcp_name"{
+    type    = string
+    default = "trendmicro_gcp"
+}
+variable "gcp_environment"{
+    type    = string
+    default = ""
+}
+variable "projectId"{
+    type    = string
+    default = ""
+}
+variable "ProjectName"{
+    type    = string
+    default = ""
+}
+variable "ServiceAccountUniqueId"{
+    type    = string
+    default = ""
+}