+++ CHANGELOG.md
@@ -44,6 +44,7 @@ FEATURES:
 * **New Data Source:** `conformity_external_id`
 * **New Resource:** `conformity_aws_account`
 * **New Resource:** `conformity_azure_account`
+* **New Resource:** `conformity_gcp_account`
 * **New Resource:** `conformity_group`
 * **New Resource:** `conformity_user`
 * **New Resource:** `conformity_sso_user`